grok/clients/python/requirements.txt
# xAI Grok — uses OpenAI-compatible SDK
openai>=1.84.0
python-dotenv>=1.1.0
httpx>=0.28.1
